+++ lib/src/platform_interface/capture_operations_interface.dart
@@ -51,6 +51,7 @@ abstract class CaptureOperationsInterface {
     int? cropY,
     int? cropWidth,
     int? cropHeight,
+    bool resizeCroppedToWindowSize = true,
   }) {
     throw UnimplementedError('captureWindow() has not been implemented.');
   }